moved filter code; adapted for arduino
